requirements.txt
Shapely==1.8.1.post1
python-frontmatter==1.0.0